Move node statistics to module Node.Statistics
