+++ src/Node/Internal.hs
@@ -22,11 +22,9 @@ module Node.Internal (
     defaultNodeEnvironment,
     nodeId,
     nodeEndPointAddress,
-    Statistics(..),
     stTotalLiveBytes,
     stRunningHandlersRemoteVariance,
     stRunningHandlersLocalVariance,
-    PeerStatistics(..),
     nodeStatistics,
     ChannelIn(..),
     ChannelOut(..),
@@ -72,6 +70,8 @@ import qualified Mockable.Metrics              as Metrics
 import qualified Network.RateLimiting          as RL
 import qualified Network.Transport             as NT (EventErrorCode (EventConnectionLost, EventEndPointFailed, EventTransportFailed))
 import qualified Network.Transport.Abstract    as NT
+import           Node.Statistics               (PeerStatistics (..), Statistics (..)
+                                               , initialStatistics)
 import           System.Random                 (Random, StdGen, random)
 import           System.Wlog                   (WithLogger, logDebug, logError, logWarning)
 import qualified Node.Message                  as Message
@@ -229,37 +229,6 @@ writeChannel (ChannelOut conn) chunks = do
       Left err -> throw err
       Right _  -> pure ()
 
--- | Statistics concerning traffic at this node.
-data Statistics m = Statistics {
-      -- | How many handlers are running right now in response to a
-      --   remotely initiated connection (whether unidirectional or
-      --   bidirectional).
-      --   NB a handler may run longer or shorter than the duration of a
-      --   connection.
-      stRunningHandlersRemote :: !(Metrics.Gauge m)
-      -- | How many handlers are running right now which were initiated
-      --   locally, i.e. corresponding to bidirectional connections.
-    , stRunningHandlersLocal :: !(Metrics.Gauge m)
-      -- | Statistics for each peer.
-    , stPeerStatistics :: !(Map NT.EndPointAddress (SharedAtomicT m PeerStatistics))
-      -- | How many peers are connected.
-    , stPeers :: !(Metrics.Gauge m)
-      -- | Average number of remotely-initiated handlers per peer.
-      --   Also track the average of the number of handlers squared, so we
-      --   can quickly compute the variance.
-    , stRunningHandlersRemoteAverage :: !(Double, Double)
-      -- | Average number of locally-initiated handlers per peer.
-      --   Also track the average of the number of handlers squared, so we
-      --   can quickly compute the variance.
-    , stRunningHandlersLocalAverage :: !(Double, Double)
-      -- | Handlers which finished normally. Distribution is on their
-      --   running time.
-    , stHandlersFinishedNormally :: !(Metrics.Distribution m)
-      -- | Handlers which finished exceptionally. Distribution is on their
-      --   running time.
-    , stHandlersFinishedExceptionally :: !(Metrics.Distribution m)
-    }
-
 stTotalLiveBytes
     :: (Mockable SharedAtomic m)
     => Statistics m -> m Int
@@ -277,20 +246,6 @@ stRunningHandlersLocalVariance :: Statistics m -> Double
 stRunningHandlersLocalVariance statistics = avg2 - (avg*avg)
     where
     (avg, avg2) = stRunningHandlersLocalAverage statistics
-
--- | Statistics about a given peer.
-data PeerStatistics = PeerStatistics {
-      -- | How many handlers are running right now in response to connections
-      --   from this peer (whether unidirectional or remotely-initiated
-      --   bidirectional).
-      pstRunningHandlersRemote :: !Int
-      -- | How many handlers are running right now for locally-iniaiated
-      --   bidirectional connections to this peer.
-    , pstRunningHandlersLocal :: !Int
-      -- | How many bytes have been received by running handlers for this
-      --   peer.
-    , pstLiveBytes :: !Int
-    }
 
 pstNull :: PeerStatistics -> Bool
 pstNull PeerStatistics{..} =
@@ -380,25 +335,6 @@ pstRemoveHandler provenance map = case provenance of
             in return $ if pstNull stats'
                         then (stats', (Map.delete peer map, True))
                         else (stats', (map, False))
-
--- | Statistics when a node is launched.
-initialStatistics :: ( Mockable Metrics.Metrics m ) => m (Statistics m)
-initialStatistics = do
-    !runningHandlersRemote <- Metrics.newGauge
-    !runningHandlersLocal <- Metrics.newGauge
-    !peers <- Metrics.newGauge
-    !handlersFinishedNormally <- Metrics.newDistribution
-    !handlersFinishedExceptionally <- Metrics.newDistribution
-    return Statistics {
-          stRunningHandlersRemote = runningHandlersRemote
-        , stRunningHandlersLocal = runningHandlersLocal
-        , stPeerStatistics = Map.empty
-        , stPeers = peers
-        , stRunningHandlersRemoteAverage = (0, 0)
-        , stRunningHandlersLocalAverage = (0, 0)
-        , stHandlersFinishedNormally = handlersFinishedNormally
-        , stHandlersFinishedExceptionally = handlersFinishedExceptionally
-        }
 
 data HandlerProvenance peerData m t =
       -- | Initiated locally, _to_ this peer. The Nonce is present if and only
